requirements.txt
Django>=5.1
djangorestframework
numpy
matplotlib
scikit-learn
django-cors-headers
reportlab
psycopg2-binary